# openapi.yaml
openapi: 3.0.3
info:
  title: ALEXZA AI API
  description: |
    ALEXZA AI provides a managed AI runtime for running actions, managing projects,
    and receiving webhooks. Use `x-api-key` for API key authentication on v1 endpoints.
  version: 1.0.0
  contact:
    name: ALEXZA AI
servers:
  - url: /
    description: API server (use full base URL in production)
...
paths:
  /v1/projects/{projectId}/run/{actionName}:
    post:
      tags: [Runtime]
      summary: Run action by name
      description: |
        Execute a saved action by project ID and action name.
        Validates input against the action's inputSchema, runs via ALEXZA Managed Runtime,
        logs usage, and deducts credits.
      operationId: runAction
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            example: "507f1f77bcf86cd799439011"
          description: MongoDB ObjectId of the project
        - name: actionName
          in: path
          required: true
          schema:
            type: string
            example: "summarize_text"
          description: Name of the action to run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunActionInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunActionSuccess"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project or action not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Runtime error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key (e.g. axza_live_...)
    SessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie (set via login)
  schemas:
    RunActionInput:
      type: object
      required: [input]
      properties:
        input:
          oneOf:
            - type: object
              description: Input object matching action's inputSchema
            - type: string
              description: For simple text input (backward compat)
      example:
        input:
          text: "Long text to summarize..."
    RunActionSuccess:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        requestId:
          type: string
        output:
          type: string
        creditsCharged:
          type: number
        usage:
          type: object
          properties:
            tokens:
              type: number
            creditsCharged:
              type: number
        latencyMs:
          type: number
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                code:
                  type: string
                message:
                  type: string
        requestId:
          type: string
          description: Present on v1 run endpoints
...